requirements.txt
# Core API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Medical Imaging Processing
pydicom==2.4.4
nibabel==5.1.0
SimpleITK==2.3.1
opencv-python==4.8.1.78
Pillow==10.1.0

# Machine Learning
torch==2.1.1
torchvision==0.16.1
torchserve==0.8.2
monai==1.3.0
scikit-learn==1.3.2
numpy==1.24.4
scipy==1.11.4

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1

# Cloud and Storage
boto3==1.34.0
botocore==1.34.0
minio==7.2.0

# Security and Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Monitoring and Logging
prometheus-client==0.19.0
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development Tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# Utilities
python-dotenv==1.0.0
click==8.1.7
tqdm==4.66.1
requests==2.31.0
